Software_requirement.txt
pip-25.0.1
streamlit-1.43.2
openai-1.66.5